Solved the valgrind error
When using the valgrind test ,invalid access occur in
the remove_head function , and another ,still reachable error
in the qtest.c ,which resulted by without free the parameter
in linenoise.c. The new version add on the function in
linenoise.h for freeing the parameter and call the function
when qtest is quit ,also modified the remove_head function
to make it robust.